Feat: 카메라 구현 완료. 세로모드 보완 필요

+++ frontend/app/src/main/java/com/ssafy/stellargram/ui/screen/cameranew/CameraXNew.kt
@@ -14,7 +14,7 @@ interface CameraXNew {
     fun getPreviewView() : PreviewView
     fun getFacingState() : StateFlow<Int>
     fun setIso(sensibility:Int)
-    fun setExposureTime(exposureTime:Int)
+    fun setExposureTime(exposureTime:Long)
     fun setClear()
 
 }
